requirements.txt
joblib==1.3.2
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
traitlets==5.14.0
tzdata==2023.4
wcwidth==0.2.12
seaborn==0.13.1